+++ R/ratio.regression.R
@@ -55,22 +55,20 @@
 ##' estimate(rmst301)
 ##' estimate(rmst302)
 ##' 
-##' ## percentage of total cumulative incidence due to cause 1
+##' ## percentage of total RMTL due to cause 1
 ##' rmtlratioI <- rmtlRatio(Event(time,cause)~platelet+tcell+age,bmt,time=30,cause=1)
 ##' summary(rmtlratioI)
 ##' 
-##' pp <- predict(rmtlratioI,bmt[1:5,])
+##' newdata <- data.frame(platelet=1,tcell=1,age=1)
+##' pp <- predict(rmtlratioI,newdata)
 ##' pp
 ##' 
-##' newdata <- data.frame(platelet=1,tcell=1,age=1)
 ##' ## percentage of total cumulative incidence due to cause 1
 ##' cifratio <- binregRatio(Event(time,cause)~platelet+tcell+age,bmt,time=30,cause=1,model="cif")
 ##' summary(cifratio)
 ##' pp <- predict(cifratio,newdata)
 ##' pp
 ##' 
-##' pp <- predict(rmtlratioI,newdata)
-##' pp
 ##' @aliases rmtlRatio
 ##' @export
 binregRatio <- function(formula,data,cause=1,time=NULL,beta=NULL,type=c("III","II","I"),
@@ -160,6 +158,7 @@ binregRatio <- function(formula,data,cause=1,time=NULL,beta=NULL,type=c("III","I
      }
   }
   Yipcw <- Y
+  cens.weights.origsort <- obs/cens.weights 
 
 obj <- function(pp,all=FALSE)
 { # {{{
@@ -354,6 +353,7 @@ hessian <- matrix(.Call("XXMatFULL",matrix(D2log,nrow=1),np,PACKAGE="mets")$XXf,
   val$model <- model[1]
   val$outcome <- outcome[1]
   val$Yipcw <- Yipcw
+  val$cenw.weights <- cens.weights.origsort 
   val$Causes <- Causes
   val$nevent <- nevent